haneul_ai/requirements.txt
# Haneul AI Agent - Core Dependencies

# FastAPI & Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# AI & LLM (GPT-5 지원)
openai>=1.40.0
anthropic>=0.35.0

# Google APIs (Gmail, Calendar)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0

# Database & Storage
sqlalchemy>=2.0.20
# sqlite3 is built-in to Python, no need to install

# File Processing & Markdown
markdown>=3.5.0
watchdog>=3.0.0
pyyaml>=6.0.1

# Security & Encryption
cryptography>=41.0.0
python-jose>=3.3.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
loguru>=0.7.2

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.5.0

# 헬스 페르소나 준비 (향후 추가)
# numpy>=2.0.0        # 운동 데이터 분석용
# pandas>=2.2.0       # 운동 기록 관리용
# matplotlib>=3.9.0   # 성과 시각화용